Properly data driven prices, embed my own jquery
