+++ .github/workflows/build.yaml
@@ -38,16 +38,13 @@ jobs:
         run: |
           npm ci
           npm run build
-          rm -rf ./_build/prices/
-          rm -rf ./_build/ych/
-          rm -rf ./_build/labs/
       - name: Setup Pages
         uses: actions/configure-pages@v5
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
           # Upload entire repository
-          path: './_build/'
+          path: './_site/'
       - name: Deploy to GitHub Pages
         id: deployment
         uses: actions/deploy-pages@v4
