+++ .gitignore
@@ -1,5 +1,5 @@
 node_modules/
-_build/
+_site/
 oldpage/
 gallery/
 Logofs.png
