+++ layouts/commissions.njk
@@ -38,16 +38,20 @@ hasBlob: true
 
     <section id="pcards">
         <h2>More examples!</h2>
-        {%- for price in collections.prices -%}
-        <div class="row" id="row{{price.data.order}}">
+
+        {% for price in prices %}
+        <div class="row">
             <!-- Image -->
-            <div class="pcard" data-bgimage="{{ price.data.imageId }}" id="{{ price.data.imageId }}" tabindex="0"></div>
+            <div class="pcard" data-bgimage="{{ price.id }}" id="{{ price.id }}" tabindex="0"></div>
             <!-- Text -->
             <div class="pcard text">
-                {{ price.templateContent | safe }}
+                <h1>{{ price.price }}€</h1>
+                <h2>{{ price.name }}</h2>
+                <h2>{{ price.description }}</h2>
             </div>
         </div>
-        {%- endfor -%}
+        {% endfor %}
+
     </section>
 
     <section class="ych-labs">
