+++ layouts/components/nav.njk
@@ -1,10 +1,10 @@
     <div class="navcont">
         <nav>
-            <div class="left">
+            <div>
                 <img src="/media/Logo.png" alt="FatDawlf's logo" id="logo">
                 <h1>FatDawlf</h1>
             </div>
-            <div class="right">
+            <div>
                 {%- if not isCharacters %}
                 <a href="/characters/">Characters</a>
                 {%- endif %}
