+++ layouts/components/ych.njk
@@ -7,16 +7,19 @@
         <a aria-roledescription="button" class="offerarrow right"> <i class="fa-solid fa-arrow-right"></i>
         </a>
         <div id="ychcards">
-            {%- for price in collections.ych -%}
+
+            {%- for price in ych -%}
                 <div class="offercard">
-                    <img src="/media/ych/{{ price.data.name }}.png"
-                    alt="{{ price.data.description }}">
+                    <img src="/media/ych/{{ price.id }}.png"
+                    alt="{{ price.description }}">
                     </img>
                     <div class="overlay">
-                        {{ price.templateContent | safe }}
+                        <h1>{{ price.name }}</h1>
+                        <h2>{{ price.price }}€</h2>
                     </div>
                 </div>
             {%- endfor -%}
+
         </div>
     </div>
 </div>
