+++ package.json
@@ -2,7 +2,9 @@
     "type": "module",
     "dependencies": {
         "@11ty/eleventy": "^3.1.2",
-        "terser": "^5.44.1"
+        "jquery": "^3.7.1",
+        "terser": "^5.44.1",
+        "yaml": "^2.8.1"
     },
     "scripts": {
         "build": "npx @11ty/eleventy",
