--- pages/labs/anim.md
@@ -1,9 +1,0 @@
----
-tags: labs
-name: animloop.avif
-description: A short breathing animation of Nathaki
----
-
-# Short animation
-
-## Ask!
