--- pages/labs/cels.md
@@ -1,9 +1,0 @@
----
-tags: labs
-name: cels.png
-description: Nathaki painted in a cel shaded style
----
-
-# Cel shaded
-
-## 40€
