--- pages/labs/refs.md
@@ -1,9 +1,0 @@
----
-tags: labs
-name: refs.png
-description: Reference sheets, starting at 40€
----
-
-# Ref. Sheets
-
-## 40€+
