--- pages/prices/1llps.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: llps
-order: 1
----
-
-# 80€
-
-## Lineless painted shaded
-
-## Detailed, lively, textured
