--- pages/prices/2lcs.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: lcs
-order: 2
----
-
-# 60€
-
-## Lined colored shaded
-
-## Refined, clean, soft
