--- pages/prices/3css.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: css
-order: 3
----
-
-# 50€
-
-## Colored Shaded Sketch
-
-## Bold, expressive, textured
