--- pages/prices/4cl.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: cl
-order: 4
----
-
-# 40€
-
-## Colored lineart
-
-## Flat, legible, simple
