--- pages/prices/5cs.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: cs
-order: 5
----
-
-# 30€
-
-## Colored sketch
-
-## Expressive, clear, focused
