--- pages/prices/6l.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: l
-order: 6
----
-
-# 25€
-
-## Lineart
-
-## Budget conscious, clean, inspiring
