--- pages/prices/7ss.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: ss
-order: 7
----
-
-# 25€
-
-## Shaded sketch
-
-## Conceptual, bold, budget conscious
