--- pages/prices/8s.md
@@ -1,11 +1,0 @@
----
-tags: prices
-imageId: s
-order: 8
----
-
-# 20€
-
-## Sketch
-
-## Budget minded, scrappy, expressive
