--- pages/ych/bellyhat.md
@@ -1,10 +1,0 @@
----
-tags: ych
-name: bellyhat
-description: Two characters, one's belly on top of the other's head.
----
-
-# Belly hat
-
-## 20/40€
-
