--- pages/ych/facesit.md
@@ -1,9 +1,0 @@
----
-tags: ych
-name: facesit
-description: Two characters, one much fatter than the other. The larger one sitting on the smaller one's face and looking quite smug.
----
-
-# Facesit
-
-## 20/40€
