--- pages/ych/gurgle.md
@@ -1,9 +1,0 @@
----
-tags: ych
-name: gurgle
-description: A character's close up image of their belly, with the text 'Gurgle' beside it.
----
-
-# Gurgle
-
-## 10€
