--- pages/ych/pinned.md
@@ -1,9 +1,0 @@
----
-tags: ych
-name: pinned
-description: A POV of someone being pinned by a larger fur.
----
-
-# Pinned!
-
-## 10/30€
