--- pages/ych/sit.md
@@ -1,10 +1,0 @@
----
-tags: ych
-name: sit
-description: Two characters, one sitting on the other's face.
----
-
-# Sit sticker
-
-## 10/20€
-
