+++ styles/nav.css
@@ -10,12 +10,13 @@
         align-items: center;
         margin: calc(var(--margin) / 1.5) calc(var(--margin) * 1.25);
         padding: .5rem 0;
+
+        & :last-child {
+            gap: calc(var(--margin) / 2);
+            margin-right: 0 !important;
+        }
     }
 
-    .right {
-        gap: calc(var(--margin) / 2);
-        margin-right: 0;
-    }
 }
 
 @media only screen and (width < 800px) and (width > 500px) {
@@ -29,15 +30,10 @@
     nav {
         margin: calc(var(--margin) / 1.5) var(--margin);
         height: calc(var(--margin) + 48px);
-    }
 
-    .right {
-        gap: calc(var(--margin) / 2);
-        margin-right: var(--nav-padding);
-    }
-
-    .left {
-        margin-left: var(--nav-padding);
+        & :last-child {
+            gap: calc(var(--margin) / 2);
+        }
     }
 }
 
@@ -51,15 +47,10 @@
     nav {
         margin: calc(var(--margin) / 1.5) var(--margin);
         height: calc(var(--margin) + 48px);
-    }
 
-    .right {
-        gap: calc(var(--margin) / 1.75);
-        margin-right: var(--nav-padding);
-    }
-
-    .left {
-        margin-left: var(--nav-padding);
+        & :last-child {
+            gap: calc(var(--margin) / 1.75);
+        }
     }
 }
 
@@ -73,15 +64,10 @@
     nav {
         margin: calc(var(--margin) / 1.5) calc(var(--margin) * 4);
         height: calc(var(--margin) + 48px);
-    }
 
-    .right {
-        gap: calc(var(--margin) / 1.5);
-        margin-right: var(--nav-padding);
-    }
-
-    .left {
-        margin-left: var(--nav-padding);
+        & :last-child {
+            gap: calc(var(--margin) / 1.5);
+        }
     }
 }
 
@@ -107,17 +93,75 @@ nav {
     transition-timing-function: var(--nav-timing);
     view-transition-name: navtransition;
 
-    translate: 0 -150%;
-    animation: appear 0.6s 1s var(--nav-timing) forwards;
-
     display: flex;
     align-items: center;
     justify-content: space-between;
-}
 
-@keyframes appear {
-    from { translate: 0 -150%; }
-    to { translate: 0; }
+    & :first-child {
+        height: inherit;
+        display: flex;
+        align-items: center;
+
+        margin-left: calc(var(--nav-padding) / 2);
+
+        &>h1 {
+            font-family: 'Nunito', sans-serif;
+            font-size: var(--nav-text-title);
+            font-weight: 1000 !important;
+            color: white;
+        }
+    }
+
+    & :last-child {
+        display: flex;
+        align-items: center;
+
+        margin-right: calc(var(--nav-padding) / 2);
+
+        &>a {
+            font-family: 'Nunito', sans-serif;
+            font-size: var(--nav-text-link);
+            font-weight: 400;
+            letter-spacing: normal;
+            text-decoration: none;
+
+
+            color: white;
+            transition-duration: 0.2s;
+            transition-delay: 0.1s;
+
+            &:hover {
+                letter-spacing: 1.5px;
+            }
+        }
+
+        & #prices {
+            padding: .5rem 1.2rem;
+            background-color: rgb(var(--button));
+            border-radius: 6rem;
+
+            font-weight: 800;
+
+            &:hover {
+                background-color: var(--main-accent);
+                font-weight: 400;
+                font-style: italic;
+            }
+        }
+
+        & #labs,
+        & #ych {
+            cursor: pointer;
+        }
+
+        & #labs:hover {
+            color: var(--labs-accent);
+        }
+
+        & #ych:hover {
+            color: var(--ych-accent);
+        }
+    }
 }
 
 ::view-transition-group(content) {
@@ -133,18 +177,6 @@ nav {
     translate: 0;
 }
 
-.left {
-    height: inherit;
-    display: flex;
-    align-items: center;
-
-    &>h1 {
-        font-family: 'Nunito', sans-serif;
-        font-size: var(--nav-text-title);
-        font-weight: 1000 !important;
-        color: white;
-    }
-}
 
 #logo {
     height: var(--logo-size);
@@ -159,43 +191,19 @@ nav {
 }
 
 @keyframes spinny {
-    from { rotate: 0; }
-    33% { rotate: -8deg; }
-    66% { rotate: 8deg; }
-    to { rotate: 0; }
-}
-
-.right {
-    display: flex;
-    align-items: center;
-
-    &>a {
-        font-family: 'Nunito', sans-serif;
-        font-size: var(--nav-text-link);
-        font-weight: 1000;
-        letter-spacing: normal;
-        text-decoration: none;
-
-        color: white;
-        transition-duration: 200ms;
-
-        &:hover {
-            letter-spacing: 1.5px;
-        }
+    from {
+        rotate: 0;
     }
 
-    & #prices {
-        padding: .5rem 1.2rem;
-        background-color: rgb(var(--button));
-        border-radius: 6rem;
-
-        &:hover {
-            background-color: var(--main-accent);
-            font-weight: 400;
-            font-style: italic;
-        }
+    33% {
+        rotate: -8deg;
     }
 
-    & #labs:hover { color: var(--labs-accent); }
-    & #ych:hover { color: var(--ych-accent); }
+    66% {
+        rotate: 8deg;
+    }
+
+    to {
+        rotate: 0;
+    }
 }
